requirements.txt
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
pillow>=9.3.0
numpy>=1.23.0
matplotlib>=3.5.0
gradio>=3.40.0
opencv-python>=4.6.0
huggingface-hub>=0.16.0